requirements.txt
streamlit
sentence-transformers
faiss-cpu
transformers
langchain
langchain-community
torch
numpy
huggingface-hub
accelerate
pandas
